requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
courts-db==0.10.27
cycler==0.12.1
eyecite==2.7.6
fast_diff_match_patch==2.1.0
fonttools==4.62.1
idna==3.11
igraph==1.0.0
joblib==1.5.3
kiwisolver==1.5.0
leidenalg==0.11.0
lxml==6.0.3
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pyahocorasick==2.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-louvain==0.16
regex==2026.4.4
reporters-db==3.2.64
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
texttable==1.7.0
threadpoolctl==3.6.0
urllib3==2.6.3
